move QgsLineString tests from QgsGeometry
